scripts/writing/pyproject.toml
[project]
name = "9seki-writing"
version = "0.1.0"
description = "LLM writing pipeline for 9seki: data → Gemini analysis → Claude draft"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.60.0",
    "google-genai>=1.0.0",
    "duckdb>=1.1.0",
    "pandas>=2.2",
    "python-slugify>=8.0",
    "pyyaml>=6.0",
]

[tool.uv]
package = false
